+++ R/accessors.R
@@ -14,7 +14,7 @@ response_vars.fbl_ts <- function(x){
   x%@%"response"
 }
 #' @export
-response_vars.mbl_df <- function(x){
+response_vars.mdl_df <- function(x){
   x%@%"response"
 }
 #' @export
@@ -35,4 +35,19 @@ distribution_var <- function(x){
 #' @export
 distribution_var.fbl_ts <- function(x){
   x%@%"dist"
+}
+
+#' Return model column variables
+#' 
+#' `mable_vars()` returns a character vector of the model variables in the
+#' object.
+#' 
+#' @param x A dataset containing models (such as a mable).
+#' @export
+mable_vars <- function(x){
+  UseMethod("mable_vars")
+}
+#' @export
+mable_vars.mdl_df <- function(x){
+  x%@%"model"
 }