+++ R/generate.R
@@ -29,7 +29,7 @@
 #' @export
 generate.mdl_df <- function(x, new_data = NULL, h = NULL, times = 1, seed = NULL, ...){
   kv <- c(key_vars(x), ".model")
-  mdls <- x%@%"model"
+  mdls <- mable_vars(x)
   if(!is.null(new_data)){
     new_data <- bind_new_data(x, new_data)[["new_data"]]
   }
@@ -68,8 +68,7 @@ generate.mdl_ts <- function(x, new_data = NULL, h = NULL, times = 1, seed = NULL
       !!!set_names(rep(list(as_tibble(new_data)), times), seq_len(times)),
       .names_to = ".rep"
     )
-    
-    new_data <- build_tsibble(new_data, index = idx, key = kv, interval = intvl)
+    new_data <- build_tsibble(new_data, index = !!idx, key = !!kv, interval = intvl)
   }
   
   # Compute specials with new_data