+++ R/reconciliation.R
@@ -67,7 +67,7 @@ forecast.lst_mint_mdl <- function(object, key_data, ...){
   if(length(unique(map(fc, interval))) > 1){
     abort("Reconciliation of temporal hierarchies is not yet supported.")
   }
-  fc_dist <- map(fc, function(x) x[[x%@%"dist"]])
+  fc_dist <- map(fc, function(x) x[[distribution_var(x)]])
   fc_mean <- as.matrix(invoke(cbind, map(fc_dist, mean)))
   fc_var <- transpose_dbl(map(fc_dist, distributional::variance))
   
@@ -80,8 +80,13 @@ forecast.lst_mint_mdl <- function(object, key_data, ...){
   }
   
   # Compute weights (sample covariance)
-  res <- map(object, function(x, ...) residuals(x, ...)[[2]], type = "response")
-  res <- matrix(invoke(c, res), ncol = length(object))
+  res <- map(object, function(x, ...) residuals(x, ...), type = "response")
+  if(length(unique(map_dbl(res, nrow))) > 1){
+    # Join residuals by index #199
+    res <- unname(as.matrix(reduce(res, full_join, by = "date")[,-1]))
+  } else {
+    res <- matrix(invoke(c, map(res, `[[`, 2)), ncol = length(object))
+  }
   
   n <- nrow(res)
   covm <- crossprod(stats::na.omit(res)) / n