+++ R/stream.R
@@ -17,7 +17,7 @@ stream <- function(object, ...){
 #' @rdname stream
 #' @export
 stream.mdl_df <- function(object, new_data, ...){
-  mdls <- object%@%"model"
+  mdls <- mable_vars(object)
   new_data <- bind_new_data(object, new_data)
   object %>% 
     dplyr::mutate_at(vars(!!!mdls),
@@ -45,6 +45,6 @@ stream.mdl_ts <- function(object, new_data, ...){
   new_data[measured_vars(object$data)] <- resp
   
   object$fit <- stream(object[["fit"]], new_data, specials = specials, ...)
-  object$data <- dplyr::bind_rows(object$data, select(new_data, !!!syms(colnames(object$data))))
+  object$data <- bind_rows(object$data, select(new_data, !!!syms(colnames(object$data))))
   object
 }